Structured Encryption/requirements.txt
absl-py==0.9.0
alabaster==0.7.12
astroid==2.4.0
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.3
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
bleach==1.5.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorama==0.4.3
crypto==1.4.1
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
docutils==0.16
entrypoints==0.3
flake8==3.8.2
Flask==0.12
future==0.18.2
gym==0.17.1
helpdev==0.7.1
html5lib==0.9999999
idna==2.9
imagesize==1.2.0
intervaltree==3.0.2
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
keyring==21.2.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
Naked==0.1.31
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.5
nltk==3.5
notebook==6.0.3
numpy==1.18.3
numpydoc==1.0.0
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
plotting==0.0.4
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
pycryptodomex==3.10.1
pydocstyle==5.0.2
pyflakes==2.1.1
pyglet==1.5.0
Pygments==2.6.1
pylint==2.5.0
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.8.0
PyQtWebEngine==5.12.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.4
QtPy==1.9.0
regex==2021.3.17
requests==2.23.0
rope==0.17.0
scikit-learn==0.24.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
shellescape==3.8.1
six==1.14.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
Sphinx==3.0.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.3
spyder-kernels==1.9.1
tensorflow-tensorboard==1.5.1
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.0
tornado==6.0.4
tqdm==4.59.0
traitlets==4.3.3
unique-id==1.0.1
urllib3==1.25.9
uuid==1.30
watchdog==0.10.2
wcwidth==0.2.2
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yapf==0.30.0
